# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitorban API",
        "description": "Monitorban REST API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "paths": {
        "/api/v1/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List server rooms",
                "description": "Display a listing of the resource.",
                "operationId": "724bfbbff387f72aa00ac8a9c5e8d9aa",
                "responses": {
                    "200": {
                        "description": "Server rooms list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a server room",
                "description": "Store a newly created resource in storage.",
                "operationId": "e2cbd0032a497fb2e481dfe1b8cd1298",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Main server room"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Server room created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/rooms/{room}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Show a server room",
                "description": "Show the specified resource.",
                "operationId": "bcf1ee58eccd7789197fe25f1492fc4a",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server room details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Server room not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a server room",
                "description": "Update the specified resource in storage.",
                "operationId": "97c0c2ef0bca70b7a9e8a97ff1574e54",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Main server room"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Server room updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Server room not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a server room",
                "description": "Remove the specified resource from storage.",
                "operationId": "d80e8529f078aeda2a02a0e0e98af9a2",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Server room deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Server room not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sensors": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "List sensors",
                "operationId": "ffe9b4bc7cdaf79939d03a3a11001618",
                "responses": {
                    "200": {
                        "description": "List of sensors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SensorData"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Create sensor",
                "operationId": "eaa7bde09cc09ba7f54e00e441a2966b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSensorData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SensorData"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sensors/{sensor}": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Show sensor",
                "operationId": "bf95671993290a7b06e8fe0dfa66215e",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SensorData"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Update sensor",
                "operationId": "24ce27fead79160091ea9735f6210de6",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSensorData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sensor updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SensorData"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Delete sensor",
                "operationId": "c44995a2a5357eac203efc5a1e8adf9e",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SensorData": {
                "description": "Sensor resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Temperature Sensor"
                    },
                    "server_room_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "type": {
                        "type": "string",
                        "example": "temperature"
                    },
                    "title_fa": {
                        "type": "string",
                        "example": "سنسور دما"
                    },
                    "title_en": {
                        "type": "string",
                        "example": "Temperature Sensor"
                    },
                    "alert_type": {
                        "type": "string",
                        "example": "warning"
                    },
                    "physical_address": {
                        "type": "string",
                        "example": "0x01",
                        "nullable": true
                    },
                    "unit": {
                        "type": "string",
                        "example": "°C",
                        "nullable": true
                    },
                    "alert_interval": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "alert_count": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "min_daily_record": {
                        "type": "integer",
                        "example": 50,
                        "nullable": true
                    },
                    "recordable_changes": {
                        "type": "string",
                        "example": "1-degree change",
                        "nullable": true
                    },
                    "has_critical_history": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_warning_history": {
                        "type": "boolean",
                        "example": false
                    },
                    "crisis_committee": {
                        "type": "boolean",
                        "example": false
                    },
                    "icon": {
                        "type": "string",
                        "example": "icon.png",
                        "nullable": true
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": "profile.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StoreSensorData": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Temp Sensor"
                    },
                    "server_room_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "type": {
                        "type": "string",
                        "example": "temperature"
                    },
                    "title_fa": {
                        "type": "string",
                        "example": "سنسور دما"
                    },
                    "title_en": {
                        "type": "string",
                        "example": "Temperature Sensor"
                    },
                    "alert_type": {
                        "type": "string",
                        "example": "warning"
                    },
                    "physical_address": {
                        "type": "string",
                        "example": "0x01",
                        "nullable": true
                    },
                    "unit": {
                        "type": "string",
                        "example": "°C",
                        "nullable": true
                    },
                    "alert_interval": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "alert_count": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "min_daily_record": {
                        "type": "integer",
                        "example": 50,
                        "nullable": true
                    },
                    "recordable_changes": {
                        "type": "string",
                        "example": "1-degree change",
                        "nullable": true
                    },
                    "has_critical_history": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_warning_history": {
                        "type": "boolean",
                        "example": false
                    },
                    "crisis_committee": {
                        "type": "boolean",
                        "example": false
                    },
                    "icon": {
                        "type": "string",
                        "example": "icon.png",
                        "nullable": true
                    },
                    "profile_picture": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateSensorData": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreSensorData"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}